chore: remove SourceSection.kt
SourceSection.kt was previously replaced by ResourceSection.kt
but, at that time, SourceSection.kt was not removed.
